Change to enum class in constitutive contact laws
cf #125

+++ src/drt_contact_constitutivelaw/coconstlaw_parameter.cpp
@@ -82,7 +82,7 @@ void CONTACT::CONSTITUTIVELAW::Container::Pack(DRT::PackBuffer& data) const
   // id_
   AddtoPack(data, id_);
   // type_
-  AddtoPack(data, type_);
+  AddtoPack(data, static_cast<int>(type_));
   // name_
   AddtoPack(data, name_);
